sandbox(debugUI): add create delete and nCreate console commands.
new commands:
- create: add a new entity.
  CREATE <STATIC/DYNAMIC> <SHAPE> <NAME>
- delete: remove entity.
  DELETE <ID>
- nCreate: batch-create N random entities.
  nCREATE <COUNT>

+++ SandBox/Core/Application.h
@@ -33,7 +33,7 @@ private :
 
 	Uint64 m_CurrTime;
 	Uint64 m_PrevTime;
-	double m_DeltaTime;
+	float m_DeltaTime;
 
 public :
